Tested RediDiscoveryService and Reformat Code
Completed the development and evaluation of basic tests for RedisDiscoveryServer. These tests Launch concurrent Regions with different ips and port but with unique RequestIdentifiers. The Players are able to discover the correct location of the other players dealing with a RequestIdentifier. Additionally there are two tests that check when a player does not send its location that the corresponding peer parties detect the failure by an exception that is thrown after some time period.

The  whole source code was reformatted with inteij beautify.