Fixed Concurrency problems due to MessageBroker and Player Interaction.
The two major problems fixed were the RequestLocks class that had an HashMap instead of a concurrent HashMap and release the locks when filteredIndex and protocolResults were read. The first problem of the RequestLocks caused some lockOnRequests methods to access a key which was being relocated and found a null value. HashMap was not appropriate for concurrent modifications and reads. The second problem was that the functions getFilteredIndexes and getProtocolResults  on ContextPlayer did not  release their corresponding locks once they have rad the values. Since until now, a player was used per request, those locks were only locked once and there was no problem. However, since in the ComposedSearchValues, the same player was used to process two protocols, than the second protocol could block, because the execution of the first protocol did not release the lock.