SafeRegions DiscoveryService integrated and tested.
This is the first commit with SafeRegions supporting a distributed HBase Cluster using a discoveryService(Redis). The network layer has been tested with concurrent requests and without concurrent requests.  Beyond the network layer, this commit has also tested batch mpc protocols.  Finally, this version tests  some parts of the HBase API, Gets, full table scans, scans, with start and stop row, and with just start row or just stop row. The HBase API is not tested with concurrent requests.

In this version of Saferegions Get and Scan requests are all converted to full table scans to force HBase to iterate over every regions on the cluster. For the three clusters to work correctly, the rows of the tables must be the same in every  cluster. Furthermore, search values and the protected ColumnFamily:Qualifier  are passed as arguments in the attributes of the scan operation, This enables the scan to search any protected column and perform most of the HBase API.