Fixed Concurrent Peer Location and Scanner Memory Leak.
When retrieving the peers location on redis there was a busy wait with a sleep inside a synchronized block which was causing problems. The code was retrofactored to fix this issue and have a much finer granularity when caching the results.

There was a memory leak detected when running tests for a long period of time. This leak happend because the ConcurrentScanCoprocessor was not closing the SecureRegionScanner which cleaned all of the messages exchanged during the request. These messages remained referenced by the MessageBroker even though they were no longer used and caused a memory leak.